Changed back to integers.

+++ src/spaceinvaders/Entities/Entity.java
@@ -10,37 +10,37 @@ package spaceinvaders.Entities;
  * @author Acer Laptop
  */
 public class Entity {
-    private float xPosition;
-    private float yPosition;
+    private int xPosition;
+    private int yPosition;
     
     public Entity() {
         xPosition = 0;
         yPosition = 0;
     }
     
-    public void setX(float newX) {
+    public void setX(int newX) {
         xPosition = newX;
     }
     
-    public void setY(float newY) {
+    public void setY(int newY) {
         yPosition = newY;
     }
     
     public void move(int newX, int newY) {
         //TODO remove this:
-        System.out.print("Moved from: " + Float.toString(xPosition) + "," + Float.toString(yPosition) + ")");
-        System.out.println("to:         " + Float.toString(xPosition + newX) + "," + Float.toString(yPosition + newY) + ")");
+        System.out.print("Moved from: " + Integer.toString(xPosition) + "," + Integer.toString(yPosition) + ")");
+        System.out.println("to:         " + Integer.toString(xPosition + newX) + "," + Integer.toString(yPosition + newY) + ")");
         
         //NOT THIS:
         xPosition += newX;
         yPosition += newY;
     }
     
-    public float getX() {
+    public int getX() {
         return xPosition;
     }
     
-    public float getY() {
+    public int getY() {
         return yPosition;
     }
 }
